refactor: HabitRecordTable 에 날짜 계산 함수 utils로 분리 그리고 주석제거
