+++ frontend/src/pages/home/dateUtils.js
@@ -1,0 +1,15 @@
+export const getCurrentWeekDates = () => {
+  const today = new Date();
+  const currentDay = today.getDay();
+
+  const daysToMonday = currentDay === 0 ? 6 : currentDay - 1;
+  const monday = new Date(today);
+  monday.setDate(today.getDate() - daysToMonday);
+  const weekDates = [];
+  for (let i = 0; i < 7; i++) {
+    const date = new Date(monday);
+    date.setDate(monday.getDate() + i);
+    weekDates.push(date.toISOString().split("T")[0]);
+  }
+  return weekDates;
+};
